+++ Model/DVBViewerClient.cs
@@ -111,7 +111,7 @@ namespace DVBViewerServerApiWrapper.Model
             var dvbApi = DVBViewerServerApi.GetCurrentInstance();
             return dvbApi?.SendApiDataAsync("dvbcommand", new List<Helper.UriParameter> {
                 new Helper.UriParameter("target", Name),
-                new Helper.UriParameter("cmd", $"-c:{channelItem.ID}")
+                new Helper.UriParameter("cmd", $"-c:{channelItem.ChannelID}")
             });
         }